refactor(cdn): drop the re-export barrel, fence the kind-dispatching pair with lint
`src/utils/cdn/index.ts` re-exported ten names straight out of
`@uploadcare/cdn-url` under a new path. That bought nothing: the library
names were already the right words, and routing them through a barrel hid
which library surface is actually in use behind an import path that looked
local. The 16 consumers now import library names from the package and ours
from the module that defines them (`utils/cdn/operations`,
`utils/cdn/origin`), so `src/utils/cdn/` holds only code we wrote.

The barrel did enforce one thing worth keeping — `parseCdnUrl` /
`serializeCdnUrl` stayed unreachable, because every bundle measured smaller
once the last consumer moved to the per-kind functions. Omission is not a
fence once the package is imported directly, so that becomes an explicit
`style/noRestrictedImports` rule naming both symbols, with the reason in
the message. Verified it fires on `parseCdnUrl` and passes `parseFileUrl`.

AGENTS.md's "src/utils/cdn is the only import path" rule is replaced by
what now holds, and picks up the `parseFileUrl`-throws rationale that lived
in the deleted docblock — including the list of boundaries that catch it.

Size-neutral: editor 49.17 -> 49.14 kB, IIFE 98.69 -> 98.55 kB brotli.
