+++ src/blocks/CloudImageEditor/src/CloudImageEditorBlock.ts
@@ -1,4 +1,5 @@
 import { ContextConsumer, ContextProvider } from '@lit/context';
+import { parseFileUrl, serializeFileUrl } from '@uploadcare/cdn-url';
 import { html, LitElement, type PropertyValues } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { createRef, ref } from 'lit/directives/ref.js';
@@ -16,7 +17,6 @@ import { ctxNameContext } from '../../../lit/ctx-name-context';
 import { LightDomMixin } from '../../../lit/LightDomMixin';
 import { RegisterableElementMixin } from '../../../lit/RegisterableElementMixin';
 import type { ConfigType, SecureDeliveryProxyUrlResolver } from '../../../types';
-import { parseFileUrl, serializeFileUrl } from '../../../utils/cdn';
 import { serializeCsv } from '../../../utils/comma-separated';
 import { debounce } from '../../../utils/debounce.js';
 import { TRANSPARENT_PIXEL_SRC } from '../../../utils/transparentPixelSrc';
